Entity relationships EVENT errors fixed

+++ src/main/java/edu/icet/repository/event/AgendaRepository.java
@@ -3,5 +3,5 @@ package edu.icet.repository.event;
 import edu.icet.entity.event.AgendaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgendaRepository extends JpaRepository<AgendaEntity,Integer> {
+public interface AgendaRepository extends JpaRepository<AgendaEntity,Long> {
 }
